+++ src/contexts/site_resource_context.rs
@@ -1,26 +1,6 @@
+use crate::serverfns::get_site::get_site;
 use lemmy_client::lemmy_api_common::site::GetSiteResponse;
-use leptos::{
-  create_blocking_resource,
-  provide_context,
-  server,
-  server_fn::codec::GetUrl,
-  Resource,
-  ServerFnError,
-};
-
-#[server(prefix = "/serverfn", input = GetUrl)]
-async fn get_site() -> Result<GetSiteResponse, ServerFnError> {
-  use crate::utils::{get_client_and_session, GetJwt};
-  use lemmy_client::LemmyRequest;
-
-  let (client, session) = get_client_and_session().await?;
-  let jwt = session.get_jwt()?;
-
-  client
-    .get_site(LemmyRequest::from_jwt(jwt))
-    .await
-    .map_err(|e| ServerFnError::ServerError(e.to_string()))
-}
+use leptos::{create_blocking_resource, provide_context, Resource, ServerFnError};
 
 pub type SiteResource = Resource<(), Result<GetSiteResponse, ServerFnError>>;
