+++ src/ui/components/common/content_actions/post_content_actions.rs
@@ -1,45 +1,9 @@
-use crate::ui::components::common::content_actions::{ContentActionType, ContentActions};
-use lemmy_client::{
-  lemmy_api_common::{
-    lemmy_db_schema::newtypes::PostId,
-    lemmy_db_views::structs::PostView,
-    post::{CreatePostReport, PostReportResponse, PostResponse, SavePost as SavePostBody},
-  },
-  LemmyRequest,
+use crate::{
+  serverfns::{report_post::ReportPost, save_post::SavePost},
+  ui::components::common::content_actions::{ContentActionType, ContentActions},
 };
+use lemmy_client::lemmy_api_common::lemmy_db_views::structs::PostView;
 use leptos::*;
-
-#[server(prefix = "/serverfn")]
-async fn save_post(id: PostId, save: bool) -> Result<PostResponse, ServerFnError> {
-  use crate::utils::{get_client_and_session, GetJwt};
-  let (client, session) = get_client_and_session().await?;
-
-  let jwt = session.get_jwt()?;
-
-  client
-    .save_post(LemmyRequest {
-      body: SavePostBody { post_id: id, save },
-      jwt,
-    })
-    .await
-    .map_err(|e| ServerFnError::ServerError(e.to_string()))
-}
-
-#[server(prefix = "/serverfn")]
-async fn report_post(post_id: PostId, reason: String) -> Result<PostReportResponse, ServerFnError> {
-  use crate::utils::{get_client_and_session, GetJwt};
-  let (client, session) = get_client_and_session().await?;
-
-  let jwt = session.get_jwt()?;
-
-  client
-    .report_post(LemmyRequest {
-      body: CreatePostReport { post_id, reason },
-      jwt,
-    })
-    .await
-    .map_err(|e| ServerFnError::ServerError(e.to_string()))
-}
 
 #[component]
 pub fn PostContentActions(
