+++ src/ui/components/common/vote_buttons.rs
@@ -1,7 +1,10 @@
 use crate::{
   contexts::site_resource_context::SiteResource,
   ui::components::common::icon::{Icon, IconType},
-  utils::{derive_user_is_logged_in, ServerAction, ServerActionFn},
+  utils::{
+    derive_user_is_logged_in,
+    types::{ServerAction, ServerActionFn},
+  },
 };
 use leptos::*;
 use leptos_router::ActionForm;
