+++ src/ui/components/layouts/base_layout/top_nav.rs
@@ -1,17 +1,14 @@
 use crate::{
-  contexts::{
-    site_resource_context::SiteResource,
-    theme_resource_context::{Theme, ThemeResource},
-  },
+  contexts::{site_resource_context::SiteResource, theme_resource_context::ThemeResource},
   i18n::*,
+  serverfns::logout::Logout,
   ui::components::common::{
     icon::{Icon, IconType},
     unpack::Unpack,
   },
-  utils::{derive_query_signal, derive_user_is_logged_in},
+  utils::{derive_query_signal, derive_user_is_logged_in, types::Theme},
 };
-use lemmy_client::LemmyRequest;
-use leptos::{server_fn::error::NoCustomError, *};
+use leptos::*;
 use leptos_router::*;
 
 #[component]
@@ -28,21 +25,6 @@ fn InstanceName() -> impl IntoView {
       </A>
     </Unpack>
   }
-}
-
-#[server(prefix = "/serverfn")]
-pub async fn logout() -> Result<(), ServerFnError> {
-  use crate::utils::{get_client_and_session, GetJwt};
-  let (client, session) = get_client_and_session().await?;
-
-  let jwt = session.get_jwt()?;
-  client
-    .logout(LemmyRequest::from_jwt(jwt))
-    .await
-    .map_err(|e| ServerFnError::<NoCustomError>::ServerError(e.to_string()))?;
-
-  session.purge();
-  Ok(())
 }
 
 #[component]
