+++ src/utils.rs
@@ -3,8 +3,7 @@ pub use derive_query_signal::*;
 
 mod derive_user_is_logged_in;
 pub use derive_user_is_logged_in::*;
-mod types;
-pub use types::*;
+pub mod types;
 
 #[cfg(feature = "ssr")]
 mod get_client_and_session;
