+++ src/utils/types.rs
@@ -1,14 +1,5 @@
-use leptos::{
-  server_fn::{client::browser::BrowserClient, codec::PostUrl, ServerFn},
-  Action,
-  ServerFnError,
-};
-use serde::de::DeserializeOwned;
-use trait_set::trait_set;
+mod theme;
+pub use theme::Theme;
 
-trait_set! {
-    pub trait ServerActionFn = DeserializeOwned + ServerFn<InputEncoding = PostUrl, Client = BrowserClient> + 'static;
-}
-
-pub type ServerAction<T> =
-  Action<T, Result<<T as ServerFn>::Output, ServerFnError<<T as ServerFn>::Error>>>;
+mod server_action;
+pub use server_action::*;
